requirements.txt
requests
python-dotenv
Pillow
python-dateutil
